Perfiles de Admi y usuarios completos(modifica contraseña y datos perosnales)

+++ TPC_Orellana_Pieruzzini/Dominio/Persona.cs
@@ -22,6 +22,6 @@ namespace Dominio
 
         public bool Estado { get; set; }
 
-        public long Cuil { get; set; }
+        public int Cuil { get; set; }
     }
 }
